+++ src/components/UXComponents/carrousel/styles/index.css
