+++ src/theme/mainTheme.js
@@ -1,0 +1,15 @@
+import {createTheme} from "@mui/material/styles";
+import colors from "../utils/colorimetría"
+
+const theme = createTheme({
+    palette: {
+        primary: {
+            main: "#210D49",
+        },
+        secondary: {
+
+        }
+    }
+})
+
+export default theme;
