+++ src/utils/colorimetría.js
@@ -1,0 +1,3 @@
+const colors = ["#210D49", "#101111", "#1CF445", "#2D302E", "#FFFFF"];
+
+export default colors;
